requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
playwright
supabase
pydantic
python-jose[cryptography]
python-multipart
passlib[bcrypt]
email-validator
pillow>=9.0.0
cairosvg>=2.5.0
